refactor: move Xcode scaffolding into services (#4902)
